requirements.txt
python-dotenv>=1.1.1
python-escpos>=3.1
